[nomad] move region files into one file for easier overview
